# Speedup UTCDateTimeAttribute deserialization