`UTCDateTimeAttribute.deserialize` spends most of it's time inside of dateutil which attempts to first
guess the format.  In most cases PynamoDB will have written the data to begin with, so we can first
optimistically assume the data is stored in DATETIME_FORMAT before falling back to dateutil.

This ~4.5 times faster than always going through dateutil.

```python
from dateutil.parser import parse
import datetime
from dateutil.tz import tzutc

DATETIME_FORMAT = '%Y-%m-%dT%H:%M:%S.%f%z'
now_str = datetime.datetime.now().replace(tzinfo=tzutc()).strftime(DATETIME_FORMAT)

%timeit parse(now_str) # 116 µs ± 9.46 µs per loop (mean ± std. dev. of 7 runs, 10000 loops each)
%timeit datetime.datetime.strptime(now_str, DATETIME_FORMAT) # 26 µs ± 7.53 µs per loop (mean ± std. dev. of 7 runs, 10000 loops each)
```